# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "aalug",
            "url": "https://github.com/aalug",
            "email": "a.a.gulczynski@gmail.com"
        }
    },
    "paths": {
...
        "/job-applications": {
            "post": {
                "description": "Create a job application. Only users can access this endpoint.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job applications"
                ],
                "summary": "Create job application",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CV file",
                        "name": "cv",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Message for the employer",
                        "name": "message",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Job ID",
                        "name": "job_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.jobApplicationResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Only users can access, not employers.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Any other error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/job-applications/employer/{id}": {
            "get": {
                "description": "Get job application for an employer. Only employers can access this endpoint. It returns different details than getJobApplicationForUser.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job applications"
                ],
                "summary": "Get job application for employer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "job application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.getJobApplicationForEmployerResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Only employers can access, not users.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Only an employer that is part of the company that created this application can access this endpoint.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Any other error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/job-applications/user/{id}": {
            "get": {
                "description": "Get job application for a user. Only users can access this endpoint. It returns different details than getJobApplicationForEmployer.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job applications"
                ],
                "summary": "Get job application for user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "job application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.getJobApplicationForUserResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Only users can access, not employers.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Only the applicant (the owner) of the job application can access this endpoint.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Any other error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "api.getJobApplicationForEmployerResponse": {
            "type": "object",
            "properties": {
                "application_date": {
                    "type": "string"
                },
                "application_id": {
                    "type": "integer"
                },
                "application_message": {
                    "type": "string"
                },
                "application_status": {
                    "$ref": "#/definitions/db.ApplicationStatus"
                },
                "job_id": {
                    "type": "integer"
                },
                "job_title": {
                    "type": "string"
                },
                "user_cv": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "user_email": {
                    "type": "string"
                },
                "user_full_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "user_location": {
                    "type": "string"
                }
            }
        },
        "api.getJobApplicationForUserResponse": {
            "type": "object",
            "properties": {
                "application_date": {
                    "type": "string"
                },
                "application_id": {
                    "type": "integer"
                },
                "application_message": {
                    "type": "string"
                },
                "application_status": {
                    "$ref": "#/definitions/db.ApplicationStatus"
                },
                "company_name": {
                    "type": "string"
                },
                "job_id": {
                    "type": "integer"
                },
                "job_title": {
                    "type": "string"
                },
                "user_cv": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "api.jobApplicationResponse": {
            "type": "object",
            "properties": {
                "applied_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "job_id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/db.ApplicationStatus"
                }
            }
        },
...
        "db.ApplicationStatus": {
            "type": "string",
            "enum": [
                "Applied",
                "Seen",
                "Interviewing",
                "Offered",
                "Rejected"
            ],
            "x-enum-varnames": [
                "ApplicationStatusApplied",
                "ApplicationStatusSeen",
                "ApplicationStatusInterviewing",
                "ApplicationStatusOffered",
                "ApplicationStatusRejected"
            ]
        },
...
    }
}